services/normalization/requirements.txt
# Normalization Service Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Models
-e ../../libs/python/ladon-models

# Google Cloud
google-cloud-pubsub==2.18.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
